chore: bump eslint-config-erb (#3412)
* chore: bump eslint-config-erb

* chore: drop husky and lint-staged

* chore: bump deps

* chore: bump all deps

* fix: bump again

* fix: bump again

* fix: check delete before delete

+++ .erb/scripts/delete-source-maps.js
@@ -1,8 +1,11 @@
+import fs from 'fs';
 import path from 'path';
 import rimraf from 'rimraf';
 import webpackPaths from '../configs/webpack.paths';
 
 export default function deleteSourceMaps() {
-  rimraf.sync(path.join(webpackPaths.distMainPath, '*.js.map'));
-  rimraf.sync(path.join(webpackPaths.distRendererPath, '*.js.map'));
+  if (fs.existsSync(webpackPaths.distMainPath))
+    rimraf.sync(path.join(webpackPaths.distMainPath, '*.js.map'));
+  if (fs.existsSync(webpackPaths.distRendererPath))
+    rimraf.sync(path.join(webpackPaths.distRendererPath, '*.js.map'));
 }
